+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## HEAD (unreleased)
 
+- Fix (`ngx-input`): textarea autosize directive respects the initial value and adjusts the height accordingly
+
 ## 44.3.0 (2023-3-21)
 
 - Feature (`ngx-large-format-dialog-content`): now accepts a `removeImageBackground` input to allow the removal of the default, white background